--- a/cmd/uplink/cmd/ls.go
+++ b/cmd/uplink/cmd/ls.go
@@ -48,6 +48,7 @@ func list(cmd *cobra.Command, args []string) error {
 		return err
 	}
 
+	// list objects
 	if len(args) > 0 {
 		src, err := fpath.New(args[0])
 		if err != nil {
@@ -77,8 +78,13 @@ func list(cmd *cobra.Command, args []string) error {
 	startAfter := ""
 	noBuckets := true
 
+	// list buckets
 	for {
-		list, err := project.ListBuckets(ctx, &storj.BucketListOptions{Direction: storj.After, Cursor: startAfter})
+		listOpts := &storj.BucketListOptions{
+			Direction: storj.Forward,
+			Cursor:    startAfter,
+		}
+		list, err := project.ListBuckets(ctx, listOpts)
 		if err != nil {
 			return err
 		}
@@ -96,7 +102,7 @@ func list(cmd *cobra.Command, args []string) error {
 		if !list.More {
 			break
 		}
-		startAfter = list.Items[len(list.Items)-1].Name
+		startAfter = list.Items[len(list.Items)-1].Name + "0"
 	}
 
 	if noBuckets {