--- a/cmd/uplink/cmd/rm.go
+++ b/cmd/uplink/cmd/rm.go
@@ -10,7 +10,6 @@ import (
 
 	"storj.io/storj/internal/fpath"
 	"storj.io/storj/pkg/process"
-	"storj.io/storj/uplink/setup"
 )
 
 func init() {
@@ -37,19 +36,12 @@ func deleteObject(cmd *cobra.Command, args []string) error {
 		return fmt.Errorf("No bucket specified, use format sj://bucket/")
 	}
 
-	access, err := setup.LoadEncryptionAccess(ctx, cfg.Enc)
-	if err != nil {
-		return err
-	}
-
-	project, bucket, err := cfg.GetProjectAndBucket(ctx, dst.Bucket(), access)
+	project, err := cfg.GetProject(ctx)
 	if err != nil {
 		return convertError(err, dst)
 	}
 
-	defer closeProjectAndBucket(project, bucket)
-
-	err = bucket.DeleteObject(ctx, dst.Path())
+	err = project.DeleteBucket(ctx, dst.Bucket())
 	if err != nil {
 		return convertError(err, dst)
 	}