--- a/pkg/metainfo/kvmetainfo/buckets.go
+++ b/pkg/metainfo/kvmetainfo/buckets.go
@@ -4,23 +4,19 @@
 package kvmetainfo
 
 import (
-	"bytes"
 	"context"
 	"strconv"
-	"time"
 
 	"github.com/zeebo/errs"
 
 	"storj.io/storj/pkg/encryption"
-	"storj.io/storj/pkg/pb"
-	"storj.io/storj/pkg/storage/meta"
 	"storj.io/storj/pkg/storage/objects"
 	"storj.io/storj/pkg/storj"
 	"storj.io/storj/storage"
 )
 
 // CreateBucket creates a new bucket or updates and existing bucket with the specified information
-func (db *Project) CreateBucket(ctx context.Context, bucketName string, info *storj.Bucket) (bucketInfo storj.Bucket, err error) {
+func (db *Project) CreateBucket(ctx context.Context, bucketName string, info *storj.Bucket) (_ storj.Bucket, err error) {
 	defer mon.Task()(&ctx)(&err)
 
 	if bucketName == "" {
@@ -58,37 +54,20 @@ func (db *Project) CreateBucket(ctx context.Context, bucketName string, info *st
 	}
 
 	if err := validateBlockSize(info.DefaultRedundancyScheme, info.DefaultEncryptionParameters.BlockSize); err != nil {
-		return bucketInfo, err
+		return storj.Bucket{}, err
 	}
 
 	if info.PathCipher < storj.EncNull || info.PathCipher > storj.EncSecretBox {
 		return storj.Bucket{}, encryption.ErrInvalidConfig.New("encryption type %d is not supported", info.PathCipher)
 	}
 
-	r := bytes.NewReader(nil)
-	userMeta := map[string]string{
-		"attribution-to":    info.Attribution,
-		"path-enc-type":     strconv.Itoa(int(info.PathCipher)),
-		"default-seg-size":  strconv.FormatInt(info.DefaultSegmentsSize, 10),
-		"default-enc-type":  strconv.Itoa(int(info.DefaultEncryptionParameters.CipherSuite)),
-		"default-enc-blksz": strconv.FormatInt(int64(info.DefaultEncryptionParameters.BlockSize), 10),
-		"default-rs-algo":   strconv.Itoa(int(info.DefaultRedundancyScheme.Algorithm)),
-		"default-rs-sharsz": strconv.FormatInt(int64(info.DefaultRedundancyScheme.ShareSize), 10),
-		"default-rs-reqd":   strconv.Itoa(int(info.DefaultRedundancyScheme.RequiredShares)),
-		"default-rs-repair": strconv.Itoa(int(info.DefaultRedundancyScheme.RepairShares)),
-		"default-rs-optim":  strconv.Itoa(int(info.DefaultRedundancyScheme.OptimalShares)),
-		"default-rs-total":  strconv.Itoa(int(info.DefaultRedundancyScheme.TotalShares)),
-	}
-	var exp time.Time
-	m, err := db.buckets.Put(ctx, bucketName, r, pb.SerializableMeta{UserDefined: userMeta}, exp)
+	info.Name = bucketName
+	newBucket, err := db.buckets.Create(ctx, *info)
 	if err != nil {
 		return storj.Bucket{}, err
 	}
 
-	rv := *info
-	rv.Name = bucketName
-	rv.Created = m.Modified
-	return rv, nil
+	return newBucket, nil
 }
 
 // validateBlockSize confirms the encryption block size aligns with stripe size.
@@ -116,23 +95,22 @@ func (db *Project) DeleteBucket(ctx context.Context, bucketName string) (err err
 	}
 
 	err = db.buckets.Delete(ctx, bucketName)
-
-	if storage.ErrKeyNotFound.Has(err) {
-		err = storj.ErrBucketNotFound.Wrap(err)
+	if err != nil {
+		return err
 	}
 
 	return err
 }
 
 // GetBucket gets bucket information
-func (db *Project) GetBucket(ctx context.Context, bucketName string) (bucketInfo storj.Bucket, err error) {
+func (db *Project) GetBucket(ctx context.Context, bucketName string) (_ storj.Bucket, err error) {
 	defer mon.Task()(&ctx)(&err)
 
 	if bucketName == "" {
 		return storj.Bucket{}, storj.ErrNoBucket.New("")
 	}
 
-	objMeta, err := db.buckets.Meta(ctx, bucketName)
+	bucket, err := db.buckets.Get(ctx, bucketName)
 	if err != nil {
 		if storage.ErrKeyNotFound.Has(err) {
 			err = storj.ErrBucketNotFound.Wrap(err)
@@ -140,58 +118,18 @@ func (db *Project) GetBucket(ctx context.Context, bucketName string) (bucketInfo
 		return storj.Bucket{}, err
 	}
 
-	return bucketFromMeta(ctx, bucketName, objMeta)
+	return bucket, err
 }
 
 // ListBuckets lists buckets
-func (db *Project) ListBuckets(ctx context.Context, options storj.BucketListOptions) (list storj.BucketList, err error) {
+func (db *Project) ListBuckets(ctx context.Context, listOpts storj.BucketListOptions) (_ storj.BucketList, err error) {
 	defer mon.Task()(&ctx)(&err)
-
-	var startAfter, endBefore string
-	switch options.Direction {
-	case storj.Before:
-		// before lists backwards from cursor, without cursor
-		endBefore = options.Cursor
-	case storj.Backward:
-		// backward lists backwards from cursor, including cursor
-		endBefore = keyAfter(options.Cursor)
-	case storj.Forward:
-		// forward lists forwards from cursor, including cursor
-		startAfter = keyBefore(options.Cursor)
-	case storj.After:
-		// after lists forwards from cursor, without cursor
-		startAfter = options.Cursor
-	default:
-		return storj.BucketList{}, errClass.New("invalid direction %d", options.Direction)
-	}
-
-	// TODO: remove this hack-fix of specifying the last key
-	if options.Cursor == "" && (options.Direction == storj.Before || options.Direction == storj.Backward) {
-		endBefore = "\x7f\x7f\x7f\x7f\x7f\x7f\x7f"
-	}
-
-	objItems, more, err := db.buckets.List(ctx, "", startAfter, endBefore, false, options.Limit, meta.Modified)
+	bucketList, err := db.buckets.List(ctx, listOpts)
 	if err != nil {
 		return storj.BucketList{}, err
 	}
 
-	list = storj.BucketList{
-		More:  more,
-		Items: make([]storj.Bucket, 0, len(objItems)),
-	}
-
-	for _, itm := range objItems {
-		if itm.IsPrefix {
-			continue
-		}
-		m, err := bucketFromMeta(ctx, itm.Path, itm.Meta)
-		if err != nil {
-			return storj.BucketList{}, err
-		}
-		list.Items = append(list.Items, m)
-	}
-
-	return list, nil
+	return bucketList, nil
 }
 
 func bucketFromMeta(ctx context.Context, bucketName string, m objects.Meta) (out storj.Bucket, err error) {