--- a/pkg/storj/bucket.go
+++ b/pkg/storj/bucket.go
@@ -26,7 +26,6 @@ type Bucket struct {
 	ID                          uuid.UUID
 	Name                        string
 	ProjectID                   uuid.UUID
-	Attribution                 string
 	Created                     time.Time
 	PathCipher                  CipherSuite
 	DefaultSegmentsSize         int64