… lib/uplink/bucket_attrs_test.go
… lib/uplink/project_test.go
… pkg/metainfo/kvmetainfo/buckets_test.go
… pkg/pb/metainfo.pb.go
… pkg/pb/pointerdb.pb.go
… satellite/metainfo/db.go
… satellite/metainfo/db_test.go
… satellite/metainfo/metainfo.go
… satellite/metainfo/service.go
… satellite/satellitedb/buckets.go
… satellite/satellitedb/dbx/satellitedb.dbx
… satellite/satellitedb/dbx/satellitedb.dbx.go
… satellite/satellitedb/locked.go
… uplink/metainfo/client.go