What: Why:
This PR is one part of the [ticket](https://storjlabs.atlassian.net/browse/V3-1967) to move bucket metadata into the satellite masterDB. I'm breaking this ticket into many small PRs so they don't get too big.
We want to move bucket metadata out of the pointerDB since its been a source of bugs and more complex code treating buckets as objects. This PR updates the uplink to use the new bucket RPCs for metainfo. Changes include:
- add bucket operations to metainfo client code
- adds a bucketStore for kvmetainfo.Projects to interact with metainfo client new bucket RPCs
- update to lib/uplink project to use bucketStore
- update packages that use lib/uplink library, this includes cmd/uplink, mobile/uplink, libuplinkc, and gateway
- remove attribute ID from bucket objects since the attribution table knows which bucket it belongs to

Please describe the tests:
 - Updated existing tests to conform to these PR changes
- Add a test to confirm bucket list pagination functionality
 
Please describe the performance impact:
I ran a couple uplink commands against storj-sim and printed debug monkit traces. The traces showed the following:
Creating buckets is faster:
- Old `mb` ~38ms-50ms
- New `mb` ~16ms

Listing buckets is about the same performance. To list one bucket
- Old `ls` ~14ms
- New `ls` ~14ms

Deleting buckets is faster:
- Old `rb` ~28ms
- New `rb` ~16ms

## Code Review Checklist (to be filled out by reviewer)
 - [ ] Does the PR describe what changes are being made?
 - [ ] Does the PR describe why the changes are being made?
 - [ ] Does the code follow [our style guide](https://github.com/storj/docs/blob/master/code/Style.md)?
 - [ ] Does the code follow [our testing guide](https://github.com/storj/docs/blob/master/code/Testing.md)?
 - [ ] Is the PR appropriately sized? (If it could be broken into smaller PRs it should be)
 - [ ] Does the new code have enough tests? (*every* PR should have tests or justification otherwise. Bug-fix PRs especially)
 - [ ] Does the new code have enough documentation that answers "how do I use it?" and "what does it do?"? (both source documentation and [higher level](https://github.com/storj/docs), diagrams?)
 - [ ] Does any documentation need updating?
 - [ ] Do the database access patterns make sense?